# fix: improve state serialization