--- a/packages/state-transition/test/perf/util/serializeState.test.ts
+++ b/packages/state-transition/test/perf/util/serializeState.test.ts
@@ -1,6 +1,6 @@
 import {itBench, setBenchOpts} from "@dapplion/benchmark";
 import {ssz} from "@lodestar/types";
-import {generatePerfTestCachedStateAltair} from "../util.js";
+import {cachedStateAltairPopulateCaches, generatePerfTestCachedStateAltair} from "../util.js";
 
 /**
  * This shows different statistics between allocating memory once vs every time.
@@ -13,8 +13,9 @@ describe.skip("serialize state and validators", function () {
     // increasing this may have different statistics due to gc time
     minMs: 60_000,
   });
-  const valicatorCount = 1_500_000;
-  const seedState = generatePerfTestCachedStateAltair({vc: 1_500_000, goBackOneSlot: false});
+  // change to 170_000 for holesky size
+  const valicatorCount = 20_000;
+  const seedState = generatePerfTestCachedStateAltair({vc: valicatorCount, goBackOneSlot: false});
 
   /**
    * Allocate memory every time, on a Mac M1:
@@ -32,6 +33,17 @@ describe.skip("serialize state and validators", function () {
     fn: () => {
       stateBytes.fill(0);
       stateType.tree_serializeToBytes({uint8Array: stateBytes, dataView: stateDataView}, 0, rootNode);
+    },
+  });
+
+  // now cache nodes
+  cachedStateAltairPopulateCaches(seedState);
+
+  itBench({
+    id: `serialize state ${valicatorCount} validators using new serializeToBytes() method`,
+    fn: () => {
+      stateBytes.fill(0);
+      seedState.serializeToBytes({uint8Array: stateBytes, dataView: stateDataView}, 0);
     },
   });
 
@@ -80,7 +92,7 @@ describe.skip("serialize state and validators", function () {
    * this is 3x faster than the previous approach.
    */
   const NUMBER_2_POW_32 = 2 ** 32;
-  const output = new Uint8Array(121 * 1_500_000);
+  const output = new Uint8Array(121 * valicatorCount);
   const dataView = new DataView(output.buffer, output.byteOffset, output.byteLength);
   // this caches validators nodes which is what happen after we run a state transition
   const validators = seedState.validators.getAllReadonlyValues();
@@ -117,6 +129,13 @@ describe.skip("serialize state and validators", function () {
         dataView.setUint32(offset, (withdrawableEpoch / NUMBER_2_POW_32) & 0xffffffff, true);
         offset += 4;
       }
+    },
+  });
+
+  itBench({
+    id: `serialize ${valicatorCount} validators from state `,
+    fn: () => {
+      seedState.validators.serializeToBytes({uint8Array: output, dataView}, 0);
     },
   });
 });