**Motivation**

State serialization happens once per epoch in n-historical state so its performance is important

**Description**

- New validator type to optimize its `value_serializeToBytes` performance, this helps make `state.validators.serialize()` 3.4x faster

before
```
    ✔ serialize 20000 validators manually                                 773.4012 ops/s    1.292990 ms/op        -      45968 runs   60.0 s
    ✔ serialize 20000 validators from state                               206.3978 ops/s    4.845013 ms/op        -      12271 runs   60.0 s
```

after
```
    ✔ serialize 20000 validators manually                                 768.2498 ops/s    1.301660 ms/op        -      45667 runs   60.0 s
    ✔ serialize 20000 validators from state                               694.0964 ops/s    1.440722 ms/op        -      41260 runs   60.0 s
```

- use the new `serializeToBytes()` api of ssz v0.15.1

part of #5968